refactor: consolidate env config into single Env.ts file
